Ensure string_writer has an exponential growth Define write for string writer, for a single character

+++ inc/goldfish/base64_stream.h
@@ -177,7 +177,7 @@ namespace goldfish { namespace stream
 			}
 
 			std::copy(data.begin(), data.end(), m_pending_encoding.begin());
-			m_cb_pending_encoding = static_cast<uint32_t>(data.size());
+			m_cb_pending_encoding = static_cast<uint8_t>(data.size());
 		}
 		auto flush()
 		{
